Add PointerValueWalker for walking pointer values
This adds the PointerValueWalker interface, which supplies the Pointer
function for introspection into pointer values.

Pointer is called for all pointer values immediately before
PointerEnter. This allows the value to be handled more like a primitive
value, in addition to allowing SkipEntry to be returned to prevent the
pointer from being "entered".